rearrange: moving components to components.rs
